First refactoring attempt
